Minor updates.
Add a note to the README file, making it explicit that this is merely
a wrapper around the existing ncurses library.  Closes #4

In an attempt to fix issue #3, the wrapper no longer wraps the
KEY_EVENT key code.  Closes #3

Update the dependency on golang.org/x/sys to a more recent version.